+++ src/presentation/controllers/signup.spec.ts
@@ -1,5 +1,5 @@
 import { InvalidParamError, MissingParamError, ServerError } from "../errors";
-import { EmailValidator } from "../protocols/email-validator";
+import { EmailValidator } from "../protocols";
 import { SignUpController } from "./signup";
 
 interface SutTypes {
